Added support for central reference folder
If you specify a reference folder using -r the query type unification
will prefer the query type name in the reference folder over the names
in the current directory (specified with -d)

If new query types are found the .queryType-Files will be copied to the
reference folder. If there is a naming conflict, the new query types
will be renamed (the .queryType-file as well as the #QueryType-field in
the processed files).